Put each form in its own component. Updated logic for inputs to minimize code needed to update state.
